requirements.txt
GitPython==3.1.13
python-dotenv==0.15.0
requests==2.25.1